requirements.txt
flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary